fix: ensure max_n is int

+++ pyphylon/models.py
@@ -850,7 +850,7 @@ def _calculate_metrics(P_confusion):
     }
 
 def _check_n_neighbors(data, n_neighbors):
-    max_n = 0.5 * min(data.shape)
+    max_n = int(0.5 * min(data.shape))
 
     if n_neighbors >= max_n:
         raise ValueError(
